Nube/requirements.txt
pandas==1.5.3
numpy==1.23.5
google-cloud-storage==2.14.0
python-dateutil==2.8.2
scikit-learn==1.5.2
pyarrow==12.0.0
fsspec
gcsfs
google-cloud-pubsub